Last lesson, course finished!

+++ app/Http/Controllers/ListingController.php
@@ -57,6 +57,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created!');
@@ -65,6 +67,12 @@ class ListingController extends Controller
     // Update listing data
     public function update(Request $request, Listing $listing)
     {
+
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -88,7 +96,19 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing)
     {
+
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted!');
+    }
+
+    public function manage()
+    {
+        // dd(auth()->user()->listings->get());
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
